refactor(landing-page): extract notification app into reusable file
